Add create chapter widget.

+++ cmd/grepbookweb/static/js/editor.js
@@ -81,6 +81,17 @@ var EditorViewModel = (function() {
     _editorEl = el.querySelector(".ql-editor");
   };
 
+  evm.createChapter = function(title) {
+    m.request({
+      method: "POST",
+      url: "/summaries/" + _brm.uid() + "/chapters/",
+      data: {heading: title},
+    }).then(function(chapJSON) {
+      _brm.prependChapter(ChapterModel(chapJSON, _brm));
+      m.redraw();
+    });
+  };
+
   function getIndexOf(arr, el) {
     if (el === null) return -1;
     for (var i = 0; i < arr.length; i++) {
@@ -143,6 +154,7 @@ var Editor = {
         m(".small-12.medium-10.medium-offset-1.columns", [
               m("br"),
               m("h2", "Chapters"), 
+              m.component(CreateChapterWidget, vm),
               m(".sortable", {config: vm.dragConfig},
                 vm.chapters.map(function(chap, index) {
                   return m("div", {
@@ -194,14 +206,14 @@ var ChapterEditor = {
     };
 
     vm.config = function(el, init) {
-      if (!init) {
-        vm._editor = new Quill(el, {
-          placeholder: 'Write your chapter summary ...',
-          theme: 'snow'
-        });
-        vm._editor.on('text-change', vm.updateDelta);
-        vm._editorEl = el.querySelector(".ql-editor");
-      }
+      if (init) return;
+      el.innerHTML = vm._chap.html();
+      vm._editor = new Quill(el, {
+        placeholder: 'Write your chapter summary ...',
+        theme: 'snow'
+      });
+      vm._editor.on('text-change', vm.updateDelta);
+      vm._editorEl = el.querySelector(".ql-editor");
     };
 
     vm.getText = function() {
@@ -213,24 +225,30 @@ var ChapterEditor = {
     };
 
     vm.updateDelta = function(delta, source) {
-      vm.change = vm.delta.compose(delta);
+      vm.delta = vm.delta.compose(delta);
     };
 
     vm.delete = function() {
       vm._chap.delete();
     };
 
-    vm.onSaveClick = function() {
+    vm.save = function() {
       vm._chap.html(vm.getText());
       vm._chap.delta(vm.getDelta());
       vm._chap.save();
+    };
+
+    vm.onSaveClick = function() {
+      vm.save();
       vm.toggleEditor();
     };
 
     vm.onDeleteClick = function() {
-      vm._chap.delete();
-      vm.toggleEditor();
-      m.redraw();
+      if (confirm("Sure you want to delete this chapter?")) {
+        vm._chap.delete();
+        vm.toggleEditor();
+        m.redraw();
+      }
     };
 
     function cleanupToolbar() {
@@ -241,6 +259,14 @@ var ChapterEditor = {
       vm._editorEl = null;
     }
 
+    // Autosave
+    setInterval(function() {
+      if (vm.delta.length() > 0) {
+        vm.save();
+        vm.delta = new Delta();
+      }
+    }, 5*1000);
+
     return vm;
   },
   view: function(vm) {
@@ -249,12 +275,55 @@ var ChapterEditor = {
         m("span.draggable.handle",
           [m("i.fa.fa-ellipsis-v.grey-draggable.draggable"), m.trust("&nbsp;&nbsp;")]), 
           m("span", vm._chap.heading())),
-          [(vm.editorShown()) ? m("div", {config: vm.config, id: vm._chap.id()}, m.trust(vm._chap.html())) : m("span", {onclick: vm.toggleEditor}, m.trust(vm._chap.html()))],
+          [(vm.editorShown()) ? m("div", {config: vm.config, id: vm._chap.id()}) : m("span", {onclick: vm.toggleEditor}, m.trust(vm._chap.html()))],
       (vm.editorShown()) ? m(".chapter-footer", [
         m("a.button.primary.small", {onclick: vm.onSaveClick}, m("i.fa.fa-save"), " Save"),
         m.trust("&nbsp;"),
         m("a.button.secondary.small", {onclick: vm.onDeleteClick}, m("i.fa.fa-trash"))]): null,
       ]); 
+  }
+};
+
+var CreateChapterWidget = {
+  controller: function(evm) {
+    var vm = {};
+    vm.isEditorMode = m.prop(false);
+    vm.heading = m.prop("");
+    vm.isWarn = m.prop(false);
+    
+    vm.toggleEditorMode = function() {
+      vm.isEditorMode(!vm.isEditorMode());
+    };
+    
+    vm.showWarning = function() {
+      vm.isWarn(!vm.isWarn());
+      setTimeout(function() {
+        vm.isWarn(!vm.isWarn());
+        m.redraw();
+      }, 2000);
+    };
+
+    vm.onCreateClick = function() {
+      if (vm.heading() === "") {
+        vm.showWarning();
+        return;
+      }
+      evm.createChapter(vm.heading());
+      vm.toggleEditorMode();
+      vm.heading("");
+    };
+    return vm;
+  },
+  view: function(vm) {
+    return (vm.isEditorMode()) ?  m("div", [
+      m("input[type=text]", {placeholder: "Enter a new chapter title", value: vm.heading(), oninput: m.withAttr("value", vm.heading)}),
+      m("a.button.primary.small", {onclick: vm.onCreateClick}, m("i.fa.fa-save"), " Create"),
+      m.trust("&nbsp;"),
+      m("a. button.secondary.small", {onclick: vm.toggleEditorMode}, m("i.fa.fa-times"), " Cancel"),
+      m.trust("&nbsp;"),
+      (vm.isWarn()) ? m("span.label.alert", {style: "vertical-align: top;"}, m("i.fa.fa-times-circle-o"), " Chapter title cannot be empty!") : null,
+    ]) : m("div.create-chapter-widget", 
+        m("a", {onclick: vm.toggleEditorMode}, "+ Add New Chapter"));
   }
 };
